feat(tauri): SrcInfo型をCommandParam.tsに移動
FormElementProp.tsで定義されていたSrcInfoをモデル層のCommandParam.tsに移動。
FormElementProp.tsは後方互換のためSrcInfoを再エクスポートする。
